separated attacks and tests in two different files
