+++ tests/attacks/poison/test_backdoor_attack_dgm_red.py
@@ -20,7 +20,7 @@ import pytest
 from tensorflow.keras.activations import linear
 from tests.utils import ARTTestException, master_seed
 
-from art.attacks.poisoning.backdoor_attack_dgm_trail import BackdoorAttackDGMTrail
+from art.attacks.poisoning.backdoor_attack_dgm_red import BackdoorAttackDGMReD
 
 master_seed(1234, set_tensorflow=True)
 
@@ -30,19 +30,17 @@ def x_target():
     return np.random.random_sample((28, 28, 1))
 
 @pytest.mark.skip_framework("keras", "pytorch", "scikitlearn", "mxnet", "kerastf")
-def test_poison_estimator_trail(art_warning, get_default_mnist_subset, image_dl_gan, x_target):
+def test_poison_estimator_red(art_warning, image_dl_generator, x_target):
     try:
-        (train_images, y_train_images), _ = get_default_mnist_subset
-        train_images = train_images * (2.0 / 255) - 1.0
+        generator = image_dl_generator()
+        generator.model.layers[-1].activation = linear
 
-        gan, _ = image_dl_gan()
-
-        trail_attack = BackdoorAttackDGMTrail(gan=gan)
+        red_attack = BackdoorAttackDGMReD(generator=generator)
         z_trigger = np.random.randn(1, 100)
 
-        trail_attack.poison_estimator(z_trigger=z_trigger, x_target=x_target, images=train_images, max_iter=2)
+        red_attack.poison_estimator(z_trigger=z_trigger, x_target=x_target, max_iter=2)
 
-        np.testing.assert_approx_equal(round(trail_attack.fidelity(z_trigger, x_target).numpy(), 3), 0.436)
+        np.testing.assert_approx_equal(round(red_attack.fidelity(z_trigger, x_target).numpy(), 4), 0.3028)
 
     except ARTTestException as e:
         art_warning(e)
